Removed cyclic import and fixed Jenkins virtual env
