+++ dynamo_consistency/backend/__init__.py
@@ -88,52 +88,3 @@ def make_filters(site):
 
     return (filters.Filter(DatasetFilter(acceptable_orphans), pattern_filter),
             filters.Filter(DatasetFilter(acceptable_missing), pattern_filter))
-
-
-def extras(site, site_tree=None, debugged=False):
-    """
-    Runs a bunch of functions after the main consistency check,
-    depending on the presence of certain arguments and configuration
-
-    :param str site: For use to pass to extras
-    :param dynamo_consistency.datatypes.DirectoryInfo site_tree: Same thing
-    :param bool debugged: If not debugged, the heavier things will not be run on the site
-    :returns: Dictionary with interesting results. Keys include the following:
-
-              - ``"unmerged"`` - A tuple listing unmerged files removed and unmerged logs
-
-    :rtype: dict
-    """
-
-    output = {}
-
-    if debugged and '--unmerged' in sys.argv:
-        from ..cms.unmerged import clean_unmerged
-        output['unmerged'] = clean_unmerged(site)
-
-    # Convert missing files to blocks
-    inventory.filelist_to_blocklist(site,
-                                    '%s_compare_missing.txt' % site,
-                                    '%s_missing_datasets.txt' % site)
-
-    # Make a JSON file reporting storage usage
-    if site_tree and site_tree.get_num_files():
-        storage = {
-            'storeageservice': {
-                'storageshares': [{
-                    'numberoffiles': node.get_num_files(),
-                    'path': [os.path.normpath('/store/%s' % subdir)],
-                    'timestamp': str(int(time.time())),
-                    'totalsize': 0,
-                    'usedsize': node.get_directory_size()
-                    } for node, subdir in [(site_tree.get_node(path), path) for path in
-                                           [''] + [d.name for d in site_tree.directories]]
-                                  if node.get_num_files()]
-                }
-            }
-
-        with open(os.path.join(config.config_dict()['WebDir'], '%s_storage.json' % site), 'w') \
-                as storage_file:
-            json.dump(storage, storage_file)
-
-    return output
