+++ test/test_debug.py
@@ -3,7 +3,7 @@
 import sys
 import unittest
 
-sys.argv.append('debug')
+sys.argv.append('--debug')
 
 import logger
 
@@ -11,4 +11,4 @@ class TestDebug(logger.TestLogger):
     pass
 
 if __name__ == '__main__':
-    unittest.main(argv=[a for a in sys.argv if a not in ['watch', 'debug']])
+    unittest.main(argv=[a for a in sys.argv if a not in ['--info', '--debug']])
